+++ packages/terra-paginator/CHANGELOG.md
@@ -3,6 +3,19 @@ ChangeLog
 
 Unreleased
 ----------
+### Added
+* Added documentation on required peerDependencies
+
+### Changed
+* Update tests to use `Terra.describeViewports` for setting viewports and use `Terra.it.validatesElement`
+
+2.17.0  - (July 11, 2019)
+------------------
+### Changed
+* Minor dependency version bump
+
+2.16.0 - (June 28, 2019)
+------------------
 ### Changed
 * Removed all inline styles from Dev-site components and implemented the same styles using external css.